+++ e2e/test-env/src/client/RicosTestApp/viewerPlugins.ts
@@ -30,8 +30,10 @@ import {
 import { pluginCollapsibleList } from 'wix-rich-content-plugin-collapsible-list/viewer';
 import { pluginTable } from 'wix-rich-content-plugin-table/viewer';
 import { mockFileUploadFunc } from '../../../../../examples/main/shared/utils/fileUploadUtil';
+import { TestAppConfig } from '../../types';
+import { ViewerPlugin } from 'wix-rich-content-common';
 
-const configs = {
+const configs: TestAppConfig['pluginsConfig'] = {
   fileUpload: {
     accept: '*',
     handleFileSelection: mockFileUploadFunc,
@@ -96,7 +98,7 @@ const plugins = {
 
 const presets = createPresets(plugins);
 
-export default pluginsPreset =>
+export default (pluginsPreset: TestAppConfig['plugins']): ViewerPlugin[] =>
   pluginsPreset
     ? pluginsPreset
         .map(plugin => presets[plugin])