+++ website/sidebars.js
@@ -44,6 +44,7 @@ module.exports = {
           'content_api/TruncateContent',
           'content_api/converters',
           'content_api/extract_media',
+          'content_api/appembed_extract',
           'content_api/ContentBuilder',
           'content_api/content_extract',
           'content_api/content_modify',